backend/requirements.txt
Django>=4.2
djangorestframework
djongo
pymongo
python-dotenv
Pillow
